+++ mypyc/primitives/struct_registry.py
